use integer in slicing

+++ src/metrics/distances.py
@@ -45,7 +45,7 @@ class Distance(ABC):
 
     def get_id_ood_split(self, in_distri_percentage=.4):
         random.shuffle(self.classes)
-        id_split = len(self.classes) * in_distri_percentage
+        id_split = int(len(self.classes) * in_distri_percentage)
         return self.classes[:id_split], self.classes[id_split:]
 
     def get_feature_dict(self):
